Make content resolver methods more idiomatic
